requirements.txt
fastapi
uvicorn[standard]
pyyaml
httpx
apscheduler
